# web/src/services/code/openapi.yaml
openapi: 3.0.3
info:
  title: API Specification
  version: 1.0.0
paths:
  /api/user/dashboard:
    get:
      responses:
        "204":
          description: No Content
      tags:
      - user
  /api/user/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelsLoginReq'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsLoginRes'
          description: OK
        "409":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModelsLoginRes'
                type: array
          description: Conflict
      tags:
      - user
  /api/user/logout:
    post:
      responses:
        "204":
          description: No Content
      tags:
      - user
  /api/user/signup:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelsCreateUserReq'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsUser'
          description: OK
        "409":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModelsUser'
                type: array
          description: Conflict
      tags:
      - user
components:
  schemas:
    ModelsCreateUserReq:
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: string
      type: object
    ModelsLoginReq:
      properties:
        email:
          type: string
        password:
          type: string
      type: object
    ModelsLoginRes:
      properties:
        email:
          type: string
        token:
          type: string
      type: object
    ModelsUser:
      properties:
        createdAt:
          format: date-time
          type: string
        email:
          type: string
        id:
          type: integer
        username:
          type: string
      type: object